Add annotation level and check run conclusion
I changed the annotation level to determine from the severity and the confidence of the concrete problem.

I also added in the config file a way to define what is the default severity and confidence level Frisk cares about.

I moved the annotation function into a separate file called bandit_annotations.

Also I added the function "getConclusion" in the github_api_helper file so we can know the true conclusion of the check run.

I had to change how we require the variables "config" and "annotationsLevels" from the config file in order the unit test to work.
Because now there are two variables that's why I had to change the require function.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>