Add tests for Annotation levels and check run conclusions
I had to export the getConclusion function in order to test it.

I add the file annotation.levels.test.js to test all cases for a severity and confidence for an issue.

I created the check.run.conclusion.test.js to test the conclusion for a check run with different annotations.

In the new files mixed_levels, neutral_annotations and warning_neutral_annotations I hardcoded annotations with different annotations levels for those tests.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>