Small changes in response to the propositions
I added directory argument for the working directory of gosec. I use this argument when creating the report file path. I changed the "workingDirectory" argument with a list of the files for scan by gosec. There is possible problem if you give the same file twice in this list with files "inputFiles".
I also decide to move the gosec arguments in a single variable as was done in bandit.js.

In parse_output I changed what Eric proposed but I changed the console.err message with a different message than the Eric's proposed.

Now when I have the "directory" argument I was able to move the deletion of the gosec.json file in an "afterEach" block because gosec.json is created in the same place every time.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>