Fix for bandit annotation problem
There is a bug when accessing the severity and confidence levels on issues found by Bandit.

Also I had to change the conclusion of a test in test/index.test.js because it was wrong. On this test there are three issues with annotation levels 'failure' so the test should fail too.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>