Fix a bug when call gosec with more than one go file
There is a bug when we call gosec with multiple go files.

The reason I belive is that in the gosecArgs array when I give another array - goFiles as argument javascirpt somehow flattens the goFiles array and Gosec recives something strange as argument.

The reason I belive this is that when I concatenate the args array with the gofile array now it works with pull requests with multiple go files.

I added a test for this case.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>