Use logs for Gosec
I noticed that many times we have problems with Gosec the only information we have is this:

"ERROR event: ENONET: no such file or directory <the file> in the parse output file!"

This happens because Gosec failed to run correctly but we don't know the reason.

It's much more useful to console log the logs from the Gosec call.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>