Fix a file name
I renamed the file networking_binding_test.go to networking_binding.go because in Go if a file is with name *_test.go that means it is a test file and networking_binding_test.go is not.

That is also the reason why when we make a scan with Gosec on this folder the output says Gosec scanned two files instead of three.

I decided to test that behaviour in the "Run gosec multiple go files" test.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>