Add more info links on Bandit issues
The json output of Bandit gives us a "more_info" weblink where there is more information about the issue.
Here is an example how it will work with this format of the issues found by Bandit:
https://github.com/MVrachev/Travic-CI---Tests/pull/724/checks?check_run_id=41872903

Gosec for now doesn't support such a thing. :(

Hopefully in future they will support this. I follow this thread for this topic:
https://github.com/securego/gosec/issues/127

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>