New summary
I changed our summary. The messages now are making more sense.

This is fix for issue https://github.com/vmware/precaution/issues/64

In this summary the annotations:
- with level 'failure' are errors
- with level 'warning' are warnings
- with level 'notice' are notices

Also I added a "text" field in the output object.
There I used that Bandit provides "more info" field for its issues. In the "more info" there is a link leading to the Bandit documentation of the issue.

Sadly Gosec doesn't support that now. :(

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>