Add precaution config file
We will need one more dependency: https://github.com/nodeca/js-yaml
It is not that easy to read properly a ymal file.
If you read it with fs.readFileSync you will have the content of a file as a string, not with objects which has properties.
Many people give an advice to use this library when dealing with ymal files.

I think that we can read for "skip_linters" key in our ymal file. With skip_linters the user can say which linters he want to skip when using Precaution.
Maybe it will be a good idea to give linter specific options. Like those for Bandit:
https://github.com/PyCQA/bandit#configuration

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>