Add automatic deploy on every successful build
I encrypted the heroku api_key and I made Travis CI to upload on every successful build.

The first section of the deploy settings is to deploy to dev-precaution dyno (heroku app) on every successful build on the master branch.
The second section is to deploy to precaution dyno (heroku app) on every successful build on a commit with tag.

That way we will deploy to precaution heroku app only on releases.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>