Fix for rerequest button doesn't work + unit test for that
Closes: https://github.com/vmware/precaution/issues/80

The problem was that when the user clicked the "Re-run" button then there would be created a check run rerequest even instead of a check suite rerequest event.
I added a small code that catches that case and I created a new unit test in index.test.js and a new file check_run_rerequested.json.

The file was needed to check if we catch the check run rerequest event correctly.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>