Fix for "across the forks" PRs problems (#113)
* Fix for two bugs

One of the bugs is when you make a pull request across forks. There was an error because it searched the files from the pr in the base repository instead of the forked repository.

The second bug is when there is a pull request across forks and then when you add commit to an existing pull request we didn't catch the
pull_request.synchronize event and then Precaution will queued for ever.

NOTE: The "Re-run" button doesn't work when there is pr across forks. The problem is that GitHub API give to Precaution an empty pull request object... I think this is a bug in the GitHub API and I will contact with them.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>