Use one folder for Gosec and Bandit
We tried before to use one folder - cache/ after we figure out there that we can't because Gosec needs a full GOPATH. We stuck with using two folders cache/ for Python and cache/go/src for Go files.

I and Antoine had discussed before that we can use one folder for both Bandit and Gosec but it would be cache/go/src.
I tried this and it worked. I removed all clutter code connecting with deciding which folder to use.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>