Catch Bandit AST parsing problem
Before we didn't check if there is a problem while parsing the AST three by Bandit and because of that we had situations when there is invalid python code which has syntax or other erros and Precaution didn't recognize that.

The consequence from this was that a non secure invalid code was said to be secure even though it can contain a lot of security issues.

This issue is easily resolved when using Bandit but the situation with Gosec is little more complicated. I will try to create a patch for this there.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>