Handle AST issues with error annotation
It's a lot better if an error annotation is created for every single file which is invalid instead of throwing an error.
That way we can report mistakes from the other files and fail the check run because of the invalid files.

I also added a unit test about this case. The unit test belongs to the bandit.spawn test file because it is a test only for python invalid files.

Also, we don't need two invalid files that's why I removed the hello.world.invalid file.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>