Change in "handle package with multiple files" gosec test
I saw we had three files with three definitions of main in the test/fixtures/go/src/vulnerable_package folder. This by definition for packages in Go is wrong.

Also, I noticed after removing the main from the rand_num.go file in the same folder that this file is useless.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>