Handle invalid go files with syntax errors
From version 1.3 of Gosec we get information about the possible syntax errors found from the golang parser.
That's something important to report to the user because if there are syntax errors in a file the file won't be scanned for security vulnerabilities.

Additionally, to reporting the golang syntax errors I added a unit test about this case.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>