Download more than 30 files
In order to download all of the diff files from a pull request, we have to access the data from GitHub page by page.

Also, there is an option to customize how many results do you want per page.

I added a unit test handling the paging but I  had to change the other unit tests a little bit in order to achieve that.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>