Delete unnecessary file
The configurations provided in the first commit are great if we use it to scan the whole project witha a recursive search through the files but our case is different.

When we use the plugin we will have the exact paths to the files and thus we don't need to use tsconfig.json at all.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>