Remove ESLint plugin from Precaution
We found out as described here: https://github.com/vmware/precaution/issues/160
that TSLint has a security plugin which can scan both TypeScript and JavaScript files.

Also, the TSLint plugin is supported as opposed to the ESLint plugin.

That's why we will replace the ESLint plugin with the TSLint plugin.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>