Small changes based on feedback
I removed the default case from the switches in the tslint_annotations.js file. As a result if a linter doesn't provide confidence in its output all issues it finds will be labeled as "warnings" whatever the severity of the issue is.
That is why I had to change the tslint.report.test.js

I did some paraphrase on some comments.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>