Add support for TSLint reference links
Because TSLint doesn't provide us with reference links and doesn't have
an API for such metadata as explained by the developer of
tslint-config-security: webschik/tslint-config-security#17

All rules are documented as part of the README for
tslint-config-security on GitHub, therefore we can use this
documentation to provide reference links.

When we have documentation of JavaScript/TypeScript issues
will educate our users on how to prevent future mistakes again.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>