Add unit tests for TSLint plugin doc
It's important to keep an eye on the TSLint plugin GitHub page.
We want to be the first notified if the project is moved/removed
from its current place.

Another thing we should check is if every rule has a documentation
on the project webpage and do we point to that documentation.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>