Reorder manual deployment Go/Gosec steps
Steps 3 and 4 from the "Setting up a manual deployment"
doc should be done before step 2.
Using the current order doesn't work because the GOPATH is needed
in order to install Gosec.

Another problem is that we are downloading Gosec with "go get".
That means that we are downloading the latest commit from the
master branch which can be unstable.
It's a lot better to download a specific version.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>