Remove redundant async keyword
Nodejs is single threaded but uses an event loop
for Asynchronous calls.
It has an async/await mechanism which should be used when we have
external service or when we need to wait for a data
(and we don't have immediate access to it).

Practically that translates to the simple idea that if we are calling
one function that is async or if it returns a Promise
then we need to await the result of that function and thus use our
own async function.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>