Remove redundant second package manager - yarn
From the start of the development of Precaution we are
using two package managers - yarn and npm.
This is meaningless given that one of them is enough.
This situation brings both complexities and sometimes
build failures and other problems.

Examples when we update package.json but not yarn.lock
and that resolves in build failures:
- https://github.com/vmware/precaution/commit/6a702fdc
- https://github.com/vmware/precaution/commit/8550ef2d

It will be better to use npm because npm is distributed with Node.js -
which means that when you download Node.js,
you automatically get npm installed on your computer.

Another reason is that most of the dependencies we are adding are npm
modules and thus it makes sense to use npm to handle them.

Closes: https://github.com/vmware/precaution/issues/190

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>