Add configuration file support and exclude option
Our users need a way to configure Precaution for their needs.

The first important option to enable will be the "exclude" option.
That way the users can exclude their tests folders and files.

We decided to give globs as a way for the users to be more flexible
when they are creating exclude rules.

Because YAML allows you to use comments and it's more human readable
than JSON we decided to use .precaution.yaml as a config file.

Closes: https://github.com/vmware/precaution/issues/179

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>