Document our decision to ignore tool specific configuration files
We are opting to ignore tool specific configuration files which may exist in
the repository and only use configuration which has been set in precaution.json

This strategy allows us to expose relevant functionality common to all
supported linters in a consistent fashion. Furthermore by not exposing tool
specific configuration via Precaution we don't have to worry about migration
paths should we choose to change the underlying tools in future.

Closes: #81

Signed-off-by: Joshua Lock <jlock@vmware.com>