Fix filename extensions in config.fileExtensions to have preceding '.'
Some of file extensions were listed in the fileExtensions array with a
preceding '.' and others were not. We must be consistent in using the '.'
prefix to all file extensions in the array, otherwise we may have unexpected
behaviours such as the filterData() function in github_api_helper.js
matching all filenames that end with 'js' or 'ts'.

Signed-off-by: Joshua Lock <jlock@vmware.com>