Add Dockerfile
As we discussed before #227
there are many benefits to creating a Docker image which will be run
on Heroku. That way we would be free from the limitations provided
by Heroku and we would be able to use Go modules.
Also, if one day we decide to move Precaution on another
hosting platform it will be really easy.

We won't need the ".profile" file anymore because we would install
Gosec in our Dockerfile as all other dependencies.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>